engine_traits: Move KvEngine methods into WriteBatchExt
