+++ components/engine_rocks/src/engine.rs
@@ -6,14 +6,13 @@ use std::path::Path;
 use std::sync::Arc;
 
 use engine_traits::{
-    Error, IterOptions, Iterable, KvEngine, Mutable, Peekable, ReadOptions, Result, WriteOptions,
+    Error, IterOptions, Iterable, KvEngine, Mutable, Peekable, ReadOptions, Result,
 };
 use rocksdb::{DBIterator, Writable, DB};
 
 use crate::db_vector::RocksDBVector;
-use crate::options::{RocksReadOptions, RocksWriteOptions};
+use crate::options::{RocksReadOptions};
 use crate::util::get_cf_handle;
-use crate::write_batch::RocksWriteBatch;
 use crate::{RocksEngineIterator, RocksSnapshot};
 
 #[derive(Clone, Debug)]
@@ -52,23 +51,6 @@ impl RocksEngine {
 
 impl KvEngine for RocksEngine {
     type Snapshot = RocksSnapshot;
-    type WriteBatch = RocksWriteBatch;
-
-    fn write_opt(&self, wb: &Self::WriteBatch, opts: &WriteOptions) -> Result<()> {
-        debug_assert_eq!(wb.get_db().path(), self.0.path(), "mismatched db path");
-        let opt: RocksWriteOptions = opts.into();
-        self.0
-            .write_opt(wb.as_inner(), &opt.into_raw())
-            .map_err(Error::Engine)
-    }
-
-    fn write_batch(&self) -> Self::WriteBatch {
-        Self::WriteBatch::new(Arc::clone(&self.0))
-    }
-
-    fn write_batch_with_cap(&self, cap: usize) -> Self::WriteBatch {
-        Self::WriteBatch::with_capacity(Arc::clone(&self.0), cap)
-    }
 
     fn snapshot(&self) -> RocksSnapshot {
         RocksSnapshot::new(self.0.clone())
