+++ components/engine_rocks/src/write_batch.rs
@@ -2,10 +2,32 @@
 
 use std::sync::Arc;
 
-use engine_traits::{self, Error, Mutable, Result};
+use engine_traits::{self, Error, Mutable, Result, WriteBatchExt, WriteOptions};
 use rocksdb::{Writable, WriteBatch as RawWriteBatch, DB};
+use crate::engine::RocksEngine;
+use crate::options::RocksWriteOptions;
 
 use crate::util::get_cf_handle;
+
+impl WriteBatchExt for RocksEngine {
+    type WriteBatch = RocksWriteBatch;
+
+    fn write_opt(&self, wb: &Self::WriteBatch, opts: &WriteOptions) -> Result<()> {
+        debug_assert_eq!(wb.get_db().path(), self.as_inner().path(), "mismatched db path");
+        let opt: RocksWriteOptions = opts.into();
+        self.as_inner()
+            .write_opt(wb.as_inner(), &opt.into_raw())
+            .map_err(Error::Engine)
+    }
+
+    fn write_batch_with_cap(&self, cap: usize) -> Self::WriteBatch {
+        Self::WriteBatch::with_capacity(Arc::clone(&self.as_inner()), cap)
+    }
+
+    fn write_batch(&self) -> Self::WriteBatch {
+        Self::WriteBatch::new(Arc::clone(&self.as_inner()))
+    }
+}
 
 pub struct RocksWriteBatch {
     db: Arc<DB>,
