+++ components/engine_traits/src/engine.rs
@@ -10,9 +10,10 @@ use crate::*;
 // a trait typedef.
 
 pub trait KvEngine:
-    Peekable
+      Peekable
     + Mutable
     + Iterable
+    + WriteBatchExt
     + DBOptionsExt
     + CFHandleExt
     + ImportExt
@@ -26,14 +27,7 @@ pub trait KvEngine:
     + 'static
 {
     type Snapshot: Snapshot<Self>;
-    type WriteBatch: WriteBatch;
 
-    fn write_opt(&self, wb: &Self::WriteBatch, opts: &WriteOptions) -> Result<()>;
-    fn write(&self, wb: &Self::WriteBatch) -> Result<()> {
-        self.write_opt(wb, &WriteOptions::default())
-    }
-    fn write_batch(&self) -> Self::WriteBatch;
-    fn write_batch_with_cap(&self, cap: usize) -> Self::WriteBatch;
     fn snapshot(&self) -> Self::Snapshot;
     fn sync(&self) -> Result<()>;
